Move Arr class.
