+++ src/Illuminate/Support/Collection.php
@@ -2,7 +2,7 @@
 
 namespace Illuminate\Support;
 
-/** @deprecated You should use \Illuminate\Collect\Collection now. */
+/** @deprecated You should use \Illuminate\Collect\Collection. */
 class Collection extends \Illuminate\Collect\Collection
 {
 }
